Upgrade sbt and Plugins To 1.0
The bintray-sbt plugin has undergone a name change (to suit “sbt-“ plugin naming conventions), as well as a move to the foundweekends catch-all/shared-maintenance organization since sbt-1.0 support was added.

The sbt-release plugin has seen a couple of patch versions released which include initial support for sbt-1.0.

This commit upgrades both plugins as starts using sbt 1.0 for this project.