Upgrade Scala and Dependency Versions
Use the latest Scala and Scalatest versions.

Unfortunately, Scalatest has some slightly outdated module dependencies (a couple of patch versions old) that conflict with sbt and Scala itself.

This is easily fixed with some exclusions of scalatest’s transitive dependencies. It should be noted that this only affects test runs and not artifact building.

When upgrading the Scalatest version further, it should be checked if these exclusions are still needed by attempting a test run with out them. If there are no dependency conflicts reported, they can be safely removed.