Move Task Execution Out of Lambdas
Sbt 1.0 disallows execution of Tasks inside anonymous functions since it can lead to uncontrolled execution.

Although this is not a problem for the simple usage we have, it is [strictly disallowed by sbt 1.0][1], so in this change the execution is moved outside of the anonymous function.

Note that this restriction is [being reconsidered for sbt 1.1][2], so this code should be revisited and possibly simplified if/when that happens.

[1]: https://github.com/sbt/sbt/blob/9dfe1a840667174a49aabe0513636587dbfc0a5c/main-settings/src/main/scala/sbt/std/TaskLinterDSL.scala#L159-L167
[2]: https://github.com/sbt/sbt/issues/3266